refactor: クラス名の修正とテストの分離
