--- src/main/services/IActualBuilderService.ts
@@ -1,5 +1,0 @@
-import { EventEntry } from '@shared/data/EventEntry';
-
-export interface IActualBuilderService {
-  buildActual(start: Date, end: Date): Promise<EventEntry | null>;
-}
