+++ src/main/services/IActualPredictiveCreationService.ts
@@ -1,0 +1,5 @@
+import { EventEntry } from '@shared/data/EventEntry';
+
+export interface IActualPredictiveCreationService {
+  generatePredictedActual(start: Date, end: Date): Promise<EventEntry | null>;
+}
