+++ src/main/services/__tests__/__mocks__/ActualAutoRegistrationFinalizerMockBuilder.ts
@@ -1,14 +1,14 @@
 import { jest } from '@jest/globals';
-import { IActualAutoRegistrationFinalizerService } from '@main/services/IActualAutoRegistrationFinalizerService';
+import { IActualAutoRegistrationFinalizer } from '@main/services/IActualAutoRegistrationFinalizer';
 import { EventEntry } from '@shared/data/EventEntry';
 
-export class ActualAutoRegistrationFinalizerServiceMockBuilder {
+export class ActualAutoRegistrationFinalizerMockBuilder {
   private finalizeRegistration: jest.MockedFunction<
     (mergedActuals: EventEntry[]) => Promise<void>
   > = jest.fn();
 
-  build(): IActualAutoRegistrationFinalizerService {
-    const mock: IActualAutoRegistrationFinalizerService = {
+  build(): IActualAutoRegistrationFinalizer {
+    const mock: IActualAutoRegistrationFinalizer = {
       finalizeRegistration: this.finalizeRegistration,
     };
     return mock;
